Added --ignore-alias to 'which' calls in setup.sh
On some systems, rm has a bash alias, often something like:
alias rm='rm -i'
/usr/bin/rm

This causes the setup.sh to not only fail to rm the file it intended, but also corrupt the alias that was present to begin with, causing rm to fail until the alias is fixed or the sheel is restarted. Which has a --ignore-alias feature which will fix the issue.