Do not call chpwd hooks in setup.zsh
By default, cd'ing on zsh calls the shell function chpwd. This is commonly used for things like showing the current directory in the title terminal by printing a string with some escape codes, but this breaks _CATKIN_SETUP_DIR. `-q` disables the chpwd hooks.

The zshbuiltins(1) man page writes about -q: "If  the -q (quiet) option is specified, the hook function chpwd and the functions in the array chpwd_functions are not called. This is useful for calls to cd that do not change the environment seen by an interactive user."