Implement multiarch support in catkin (#545)
This implements multiarch support (#545), and supercedes
(#604). A two step approach uses gcc -print-multiarch and,
if failed, fallback to use dpkg-architecture to identify
an appropriate lib folder suffix to use for LD_LIBRARY_PATH
and PKG_CONFIG_PATH. builder.py and
catkin_generate_environment.cmake are modified
accordingly.

Thanks to @j-rivero for his contributions, which are hidden in
this squashed commit.