Add --no-push to catkin_prepare_release
The catkin_prepare_release command is useful to
format the changelogs and create the tags, but will
fail if no remote repository is configured for the current
branch.

Since I could not see a reaason for this limitation, I propose
the --no-push option, that will do everything locally and not
push anything.