Replace l:$absolute_path in pkgconfig by just the absolute path
The syntax l:$library_absolute_path seems to be unsupported by
upstream ld and it is failing in some versions of the linker.

This patch workaround about the problem by generating pkgconfig
files which does not set the -l (only works with relative paths)
and just provide the absolute path to the library. The compiler
will link the library as a shared lib (if -static was not supplied).