[indigo/jade] Dynamically check gtest library type
From the CMake docs: https://cmake.org/Wiki/CMake/Tutorials/Exporting_and_Importing_Targets

"
In some cases, such as when using the result of a find_library command to create an imported target, the type of the library (static or shared) is not known. As of CMake 2.6.2 one may specify this explicitly:
```
 find_library(FOO_LIBRARY NAMES foo)
 add_library(foo UNKNOWN IMPORTED)
 set_property(TARGET foo PROPERTY IMPORTED_LOCATION "${FOO_LIBRARY}")
 add_executable(myexe src1.c src2.c)
 target_link_libraries(myexe foo)
```
"

It looks like catkin requires `cmake_minimum_required(VERSION 2.8.3)` so the `UNKNOWN` keyword is supported. Tested without a pre-built (the most common catkin case), and with a shared/static prebuilt.