Ignore _CATKIN_SETUP_DIR in devel_isolated setup scripts
Due to the chaining of the setup scripts, the `_CATKIN_SETUP_DIR`
variable doesn't act the same in `devel_isolated` as it does with
`install_isolated` or a non-isolated `devel` space.

We could feasibly make it act in the same way, but since
`devel_isolated` isn't relocatable to begin with, it would have no
benefit.