Add error message if devel_isolated relocation is attempted
A devel_isolated space cannot be relocated (moved to a path other than
the one at which it was built). To use the `setup.sh` of a relocated
space, the `_CATKIN_SETUP_DIR` variable needs to be set to the new
location because the script can't reliably detect the location on its
own.

If the `_CATKIN_SETUP_DIR` is set prior to invoking
`devel_isolated/setup.sh` and the value differs from the path at which
the workspace was built, produce an error message.

At present, other shells do not use the value of `_CATKIN_SETUP_DIR` set
prior to invocation.