Correctly set gtest 1.8 root directory
In gtest version 1.8 or higher, there is a CMakeLists.txt in the
parent directory of the gtest source directory which we need to find.
This can be either /usr/src/googletest or
${CMAKE_SOURCE_DIR}/googletest.

The current version would only check in /usr/src/googletest even if
gtest was found in the project's src/ directory.

This would break the build on Ubuntu 18.04, where the system install of
gmock is version 1.8, if your project includes a newer version of gtest
in its src/ directory. The result would be that headers from src/ would
be used, but the test executables would be linked to the binaries from
/usr/src.