Fix CATKIN_SYMLINK_INSTALL with add_subdirectory()
catkin_symlink_install_files/directory/programs() use generated
functions that store the value of CMAKE_CURRENT_SOURCE_DIR at the time
they are generated. This value is incorrect when install() rules are
called from a directory using `add_subdirectory()` as is the case in
gencpp. This fixes the issue by passing the current source directory to
the generated functions at the time that they're called.

Signed-off-by: Shane Loretz<sloretz@openrobotics.org>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>