Port symlink_install fixes from ament_cmake to catkin
Symlink installs in catkin are used by colcon to provide a similar
experience to a catkin devel space. To be clear, this code is not used
by catkin itself or catkin_tools.

Specifically, this change includes fixes from the following ament_cmake
commits:
* 8b92e4affbd18b08a703897698960007738ee1b5
* fb1eda91c16a9423a8a96541e878358289e1b2b2
* 69afa32843a95dbed072c3eee282424a214f878a

In my testing, these fixes were necessary to support symlink builds of
the vast majority of ROS 1 packages released to date. Without this
change, there are very common packages in ROS Noetic today which cannot
be built with colcon when catkin symlinking is enabled.